Updated libraries (new AS, Gradle...) Added logic for multiple user push notifications. Added dynamic notification loading, and scroll dynamic loading

+++ app/src/main/java/io/relayr/tellmewhen/app/RuleValueCreateFragment.java
@@ -34,7 +34,7 @@ public class RuleValueCreateFragment extends WhatFragment {
 
         view.setOnDoneClickListener(new RuleValueView.OnDoneClickListener() {
             @Override
-            public void onDoneClicked(int value, OperatorType mCurrentOperator) {
+            public void onDoneClicked(float value, OperatorType mCurrentOperator) {
                 Storage.getRule().value = value;
                 Storage.getRule().operatorType = mCurrentOperator.ordinal();
