+++ app/src/main/java/io/relayr/tellmewhen/app/RuleValueEditFragment.java
@@ -35,7 +35,7 @@ public class RuleValueEditFragment extends WhatFragment {
 
         view.setOnDoneClickListener(new RuleValueView.OnDoneClickListener() {
             @Override
-            public void onDoneClicked(int value, OperatorType mCurrentOperator) {
+            public void onDoneClicked(float value, OperatorType mCurrentOperator) {
                 Storage.getRule().value = value;
                 Storage.getRule().operatorType = mCurrentOperator.ordinal();
